refactor CrystalGraphConverter into its own file
