+++ chgnet/graph/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from chgnet.graph.crystalgraph import CrystalGraph, CrystalGraphConverter, BatchedGraph
-from chgnet.graph.graph import Graph, Node
+from chgnet.graph.crystalgraph import CrystalGraph, BatchedGraph
+from chgnet.graph.converter import CrystalGraphConverter
 
-__all__ = ["CrystalGraph", "CrystalGraphConverter", "BatchedGraph", "Graph", "Node"]
+__all__ = ["CrystalGraph", "BatchedGraph", "CrystalGraphConverter"]
